Fix for issue #54

+++ tools/c3s/c3s.sh
@@ -7,9 +7,14 @@ if [ -f download.tar.gz ]; then
    cdo -f nc -t ecmwf copy tmpg.grib tmp.nc
 elif [ -f download.zip ]; then
    unzip download.zip
-   cat *.grib > tmp.grib
-   cdo setgridtype,regular tmp.grib tmpg.grib
-   cdo -f nc -t ecmwf copy tmpg.grib tmp.nc
+   files=(*.nc)
+   if [ -e "${files[0]}" ]; then
+     mv ${files[0]} tmp.nc
+   else
+     cat *.grib > tmp.grib
+     cdo setgridtype,regular tmp.grib tmpg.grib
+     cdo -f nc -t ecmwf copy tmpg.grib tmp.nc
+   fi
 elif [ -f download.grib ]; then
    cdo -f nc -t ecmwf copy download.grib tmp.nc
 elif [ -f download.nc ]; then
